requirements.txt
djangorestframework
psycopg2-binary
python-decouple
djangorestframework-simplejwt
pillow
django-filter
drf-yasg
python-slugify
redis
celery
django-cors-headers
python-telegram-bot